Change `nil` request body to use a configured `outgoing_model_class`
The `nil` approach changes the behavior when supplying a model you might
not know that it's `nil`. Before it raised, the new behavior is that it no
longer raises, but sends an empty body to the API.

`outgoing_model_class` is an optional argument which defaults to `nil`.
We therefore cannot use `nil` to specify that the body should be empty. We
use `:empty` instead to denote that the request should send an empty body.