Remove `StandardError` from retry mechanism
The goal of the retry mechanism is to retry on network errors. If we
encounter a `StandardError`, that is most likely not a network error and
we should not retry it. It could be a bug in ioki-ruby, or incorrect
input from the calling application. Either way, a retry is unlikely to
fix the situation.