Fix retry of `index` endpoints with `auto_paginate: true`
When during an index auto_paginate a request failed, the retry mechanism
started back at the first page. If a block was passed, that block was
called again with all items of the previously already loaded pages. This
caused blocks to be called multiple times and could result in duplicate
data for the caller.

This fixes that issue by moving the retry mechanism directly to the HTTP
request and ensures that only the request is retried and not any other
logic around it.