OpenAPI repair tool (#287)
* Load rake tasks from lib/tasks

* Add all missing schema_path / unvalidated to all models

* Initial version of the ioki:openapi:fix rake task

* Create OpenApi module where specifications and model classes are defined

* Repair deprecated attributes as well

* Adds operator API to OpenApi specifications

* Add rake task to report errors as YAML

* Fix rubocop issues

* Driver Receipt schema has been added to drive openapi json

* Skip repair attempt when no schema can be found for the model

* Use updated API schema names

* Remove spec/helper/openapi_matcher.rb again

* Fix OpenApi module to work with nested module namespaces

* Update final missing schema_paths

* Delete invalid operator models

These models are not used in the codebase and
are not present in the operator API openAPI json schema.

They are the result of a sloppy copy and paste
from the platform API models in commit:

f17abc3e4b2b57db1f43d6c8efa1501708dcb96a

* Remove Ioki::Model::Operator::NotificationSettings

This models is not used in the codebase and
is not present in the operator API openAPI json schema.

It might have been used in the User model but that class now
uses an array of the singular NotificationSetting class.

* Add a test for OpenApi::Model#repair

* Enhance OpenAPI::Model#repair to be able to deprecate attributes

* Work with multiple schemas from the OpenAPI definitions

* Remove schema_paths with _schema suffix

* Deal with oneOf and anyOf correctly

* Fix OpenApi test to pass when fixtures are missing

* Silence two necessary rubocop issues